refactor(global): adopted clearer naming convention
