+++ lib/authentication_with_google/application/google_auth_bloc/google_sign_in_bloc.dart
@@ -1,7 +1,7 @@
 import 'package:bloc/bloc.dart';
+import 'package:fpb/authentication_with_google/domain/i_google_repository_facade.dart';
 import 'package:fpb/core/domain/user.dart';
 import 'package:fpb/core/failures/auth_failure.dart';
-import 'package:fpb/sign_in_with_google/domain/i_google_repository_facade.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:injectable/injectable.dart';
