+++ lib/authentication_with_google/application/google_auth_bloc/google_sign_in_state.dart
