+++ lib/authentication_with_google/view/loading_indicator.dart
