+++ lib/senders_profile/view/senders_profile_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:fpb/assets/fpb_icons/fpb_icons_icons.dart';
 import 'package:fpb/core/shared/presentation/theming/colors/colors.dart';
-import 'package:fpb/home_screen/view/home_screen.dart';
+import 'package:fpb/home/view/home_screen.dart';
 import 'package:fpb/l10n/l10n.dart';
 
 class SendersProfilePage extends StatelessWidget {
