+++ lib/sign_up/sign_up.dart
@@ -1,0 +1,1 @@
+export 'view/signup_page.dart';
