+++ widgetbook/home_page/home_page.dart
@@ -1,5 +1,5 @@
 import 'package:fpb/core/domain/user.dart';
-import 'package:fpb/home_screen/home_screen.dart';
+import 'package:fpb/home/home_screen.dart';
 import 'package:widgetbook/widgetbook.dart';
 
 WidgetbookUseCase get homePageUsecase => WidgetbookUseCase(
